requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altgraph==0.17.4
anyio==4.10.0
APScheduler==3.11.0
attrs==25.3.0
beautifulsoup4==4.13.5
bilibili-api-python==17.3.0
Brotli==1.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
colorama==0.4.6
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.1
darkdetect==0.8.0
fonttools==4.60.0
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jieba==0.42.1
kiwisolver==1.4.9
lxml==5.4.0
matplotlib==3.10.6
multidict==6.6.4
numpy==2.3.3
packaging==25.0
pandas==2.3.2
pefile==2023.2.7
pillow==11.2.1
propcache==0.3.2
pycparser==2.23
pycryptodomex==3.23.0
pygame==2.6.1
PyJWT==2.10.1
pyparsing==3.2.4
PyQt-Fluent-Widgets==1.8.7
PyQt5==5.15.11
PyQt5-Frameless-Window==0.7.3
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.2
qrcode==8.2
qrcode-terminal==0.8
requests==2.32.5
six==1.17.0
sniffio==1.3.1
snownlp==0.12.3
soupsieve==2.8
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
wordcloud==1.9.4
yarl==1.20.1
